update class role

+++ MainProjectNew/MainProjectNew/MainProjectNew/Controllers/ClassesController.cs
@@ -13,32 +13,51 @@ namespace MainProjectNew.Controllers
     public class ClassesController : Controller
     {
         private ModelMain db = new ModelMain();
-        [Authorize(Roles = "Admin, Staff, Manager")]
         // GET: Classes
         public ActionResult Index()
         {
-            return View("Index", "_Layout",db.Classes.ToList());
+            if (User.IsInRole("Admin") || User.IsInRole("Manager") || User.IsInRole("Staff") )
+            {
+                return View("Index", "_Layout", db.Classes.ToList());
+            }
+            else
+            {
+                return RedirectToAction("Contact", "Home");
+            }
         }
-        [Authorize(Roles = "Admin, Staff, Manager")]
         // GET: Classes/Details/5
         public ActionResult Details(string id)
         {
-            if (id == null)
+            if (User.IsInRole("Admin") || User.IsInRole("Manager") || User.IsInRole("Staff") )
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Class @class = db.Classes.Find(id);
+                if (@class == null)
+                {
+                    return HttpNotFound();
+                }
+                return View("Details", "_Layout", @class);
             }
-            Class @class = db.Classes.Find(id);
-            if (@class == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Contact", "Home");
             }
-            return View("Details", "_Layout",@class);
+
         }
-        [Authorize(Roles = "Admin, Staff")]
         // GET: Classes/Create
         public ActionResult Create()
         {
-            return View("Create", "_Layout");
+            if (User.IsInRole("Admin")|| User.IsInRole("Staff"))
+            {
+                return View("Create", "_Layout");
+            }
+            else
+            {
+                return RedirectToAction("Contact", "Home");
+            }
         }
 
         // POST: Classes/Create
@@ -46,32 +65,47 @@ namespace MainProjectNew.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin, Staff")]
         public ActionResult Create([Bind(Include = "ClassID,num,ClassName")] Class @class)
         {
-            if (ModelState.IsValid)
+            if (User.IsInRole("Admin") || User.IsInRole("Staff"))
             {
-                db.Classes.Add(@class);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Classes.Add(@class);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+
+                return View("Create", "_Layout", @class);
+            }
+            else
+            {
+                return RedirectToAction("Contact", "Home");
             }
 
-            return View("Create", "_Layout",@class);
         }
-        [Authorize(Roles = "Admin, Staff")]
         // GET: Classes/Edit/5
         public ActionResult Edit(string id)
         {
-            if (id == null)
+            if (User.IsInRole("Admin") || User.IsInRole("Staff"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Class @class = db.Classes.Find(id);
+                if (@class == null)
+                {
+                    return HttpNotFound();
+                }
+                return View("Edit", "_Layout", @class);
+
             }
-            Class @class = db.Classes.Find(id);
-            if (@class == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Contact", "Home");
             }
-            return View("Edit", "_Layout",@class);
         }
 
         // POST: Classes/Edit/5
@@ -79,44 +113,69 @@ namespace MainProjectNew.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin, Staff")]
         public ActionResult Edit([Bind(Include = "ClassID,num,ClassName")] Class @class)
         {
-            if (ModelState.IsValid)
+            if (User.IsInRole("Admin") || User.IsInRole("Staff"))
             {
-                db.Entry(@class).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+
+                if (ModelState.IsValid)
+                {
+                    db.Entry(@class).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View("Edit", "_Layout", @class);
             }
-            return View("Edit", "_Layout",@class);
+            else
+            {
+                return RedirectToAction("Contact", "Home");
+            }
+
         }
 
         // GET: Classes/Delete/5
-        [Authorize(Roles = "Admin, Staff")]
         public ActionResult Delete(string id)
         {
-            if (id == null)
+            if (User.IsInRole("Admin") || User.IsInRole("Staff"))
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Class @class = db.Classes.Find(id);
+                if (@class == null)
+                {
+                    return HttpNotFound();
+                }
+                return View("Delete", "_Layout", @class);
+
             }
-            Class @class = db.Classes.Find(id);
-            if (@class == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Contact", "Home");
             }
-            return View("Delete", "_Layout",@class);
+
+
         }
 
         // POST: Classes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin, Staff")]
         public ActionResult DeleteConfirmed(string id)
         {
-            Class @class = db.Classes.Find(id);
-            db.Classes.Remove(@class);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (User.IsInRole("Admin") || User.IsInRole("Staff"))
+            {
+
+                Class @class = db.Classes.Find(id);
+                db.Classes.Remove(@class);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+
+            }
+            else
+            {
+                return RedirectToAction("Contact", "Home");
+            }
         }
 
         protected override void Dispose(bool disposing)
